resolve negligible warnings

+++ mosquitto/lib/logging_mosq.c
@@ -33,7 +33,7 @@ int _mosquitto_log_printf(struct mosquitto *mosq, int priority, const char *fmt,
 
 	pthread_mutex_lock(&mosq->log_callback_mutex);
 	if(mosq->on_log){
-		len = strlen(fmt) + 500;
+		len = (int)(strlen(fmt) + 500);
 		s = _mosquitto_malloc(len*sizeof(char));
 		if(!s){
 			pthread_mutex_unlock(&mosq->log_callback_mutex);
